requirements.txt
openai
gradio
numpy
python-dotenv